- Increased complier warnings in Eclipse and adjusted code accordingly,   usless casts and elses and the like, in Fossil. - Adjusted Strata version number in pom.xml in Fossil.
git-svn-id: http://blogometer.com/svn/open/trunk/fossil@2267 3df9ba58-24f1-0310-ade5-ba95bd924a04

+++ core/src/main/java/com/agtrz/fossil/Fossil.java
@@ -68,7 +68,7 @@ implements Strata.Storage<T>, Serializable
 
         public Object getNullKey()
         {
-            return (Long) 0L;
+            return 0L;
         }
 
         public boolean isKeyNull(Object object)
